+++ Avalon/Avalon.Draw/BrushJoin.cs
@@ -2,6 +2,6 @@ namespace Avalon.Draw;
 
 public class BrushJoin : Any
 {
-    public virtual int Index { get; set; }
+    public virtual long Index { get; set; }
     internal virtual ulong Intern { get; set; }
 }
